core-data: Move setNestedValue to block editor package
