+++ packages/block-editor/src/private-apis.js
@@ -5,7 +5,7 @@ import * as globalStyles from './components/global-styles';
 import { ExperimentalBlockEditorProvider } from './components/provider';
 import { lock } from './lock-unlock';
 import { getRichTextValues } from './components/rich-text/content';
-import { kebabCase } from './utils/object';
+import { kebabCase, setNestedValue } from './utils/object';
 import ResizableBoxPopover from './components/resizable-box-popover';
 import { ComposedPrivateInserter as PrivateInserter } from './components/inserter';
 import { PrivateListView } from './components/list-view';
@@ -29,6 +29,7 @@ lock( privateApis, {
 	ExperimentalBlockEditorProvider,
 	getRichTextValues,
 	kebabCase,
+	setNestedValue,
 	PrivateInserter,
 	PrivateListView,
 	ResizableBoxPopover,
